Remove empty atlas-demo-Hyy for Binder
When bringing the repo to life with Binder
(https://mybinder.org/)
the build of the live environment failed due to error
> fatal: No url for submodule path 'atlas-demo-Hyy/...' in .gitmodules

Binder was reading the empty directory atlas-demo-Hyy as a submodule for
which there was no corresponding entry with URL in .gitmodules.
Simplest fix is to remove atlas-demo-Hyy, since it is empty anyway.