Make section on functions in Python
Introduce Python functions with as little overhead as possible.  As an
example of using functions, terms of the Fibonacci sequence are
calculated.  The terms are printed to screen to start, then are returned
as a list once the `return` statement is introduced.  As the final
example, calls to the function are made from a snippet of relatively
complicated logic.  The section ends with a look ahead to where
functions are used in the rest of the course of notebooks.